Added MapboxOfflineNavigator and removed snapshots
