+++ build.gradle
@@ -32,8 +32,6 @@ allprojects {
     jcenter()
     maven { url 'https://plugins.gradle.org/m2' }
     maven { url 'https://mapbox.bintray.com/mapbox' }
-    // TODO Remove SNAPSHOT when 4.1.0 gets released
-    maven { url "https://oss.sonatype.org/content/repositories/snapshots/" }
   }
 
   group = GROUP
