+++ gradle/dependencies.gradle
@@ -9,8 +9,7 @@ ext {
 
   version = [
       mapboxMapSdk       : '6.7.0',
-      // TODO Remove SNAPSHOT when 4.1.0 gets released - See https://github.com/mapbox/mapbox-java/pull/895
-      mapboxSdkServices  : '4.1.0-SNAPSHOT',
+      mapboxSdkServices  : '4.1.0',
       mapboxEvents       : '3.5.2',
       mapboxNavigator    : '3.4.0',
       searchSdk          : '0.1.0-SNAPSHOT',
